fix: explicit cast to uint32_t in miner.cpp to suppress UBSan warning

+++ src/node/miner.cpp
@@ -171,7 +171,7 @@ std::shared_ptr<CBlockTemplate> BlockAssembler::CreateNewBlock()
         pblock->nVersion = gArgs.GetIntArg("-blockversion", pblock->nVersion);
     }
 
-    pblock->nTime = TicksSinceEpoch<std::chrono::seconds>(NodeClock::now());
+    pblock->nTime = static_cast<uint32_t>(TicksSinceEpoch<std::chrono::seconds>(NodeClock::now())); 
     m_lock_time_cutoff = pindexPrev->GetMedianTimePast();
 
     int nPackagesSelected = 0;
